Updates links in readme
The AM2RLauncher reddit link was updated, to point to the 2.0 release, instead of the now deprecated 1.2 version.
Also updated the License and Contritbutors link to point to the actual repo, instead of pointing to the now not-existing Lojemiru repo. Technically, the links were not broken, but this just looks nicer.